Move rendering from UIApplication to UIScreen
